types split
